working 2D models and velaution

+++ build_tubes.py
@@ -16,7 +16,7 @@ import torch.utils.data as data_utils
 from modules import utils
 from modules.evaluation import evaluate_tubes
 from modules.box_utils import decode, nms
-from data import custum_collate
+from data import custum_collate, get_gt_video_list
 import modules.gen_agent_paths as gen_paths #update_agent_paths, copy_live_to_dead,
 from modules.tube_helper import trim_tubes
 
@@ -44,11 +44,11 @@ def build_tubes(args, val_dataset):
         
         tube_file = args.save_root + 'tubes.json'
 
-        
-        paths = perform_building(args, val_dataset, iteration, tube_file)
-        
+        video_list = get_gt_video_list(val_dataset.anno_file, args.test_subsets)
+
         if args.compute_tubes:
-            make_tubes(args, paths, tube_file)
+            paths = perform_building(args, video_list, iteration)
+            make_tubes(args, paths, video_list, paths, tube_file)
         
         torch.cuda.synchronize()
         print('Computation time {:0.2f}'.format(time.perf_counter() - tt0))
@@ -60,44 +60,41 @@ def build_tubes(args, val_dataset):
                 mAP, aps, ap_strs = evaluate_tubes(
                     val_dataset.anno_file, tube_file, args.all_classes[nlt], label_type, subset, iou_thresh=0.2)
                 name = subset + '-' + label_type
-                rstr = '\n\nResults for ' + name + '\n\n'
+                rstr = '\n\nResults for ' + name + '\n'
                 print(rstr)
                 log_file.write(rstr+'\n')
                 results[name] = {'mAP': mAP, 'APs': aps}
                 for ap_str in ap_strs:
                     print(ap_str)
                     log_file.write(ap_str+'\n')
+        
+            # return  aps, aps_all, ap_strs
+            with open(result_file, 'w') as f:
+                json.dump(results, f)
+
         log_file.close()
-        # return  aps, aps_all, ap_strs
-        with open(result_file, 'w') as f:
-            json.dump(results, f)
 
-
-def perform_building(args, video_list, iteration, tube_file):
+def perform_building(args, video_list, iteration):
 
     """Test a FPN network on an image database."""
     
     # opts
-    jumpgap = 2
-    iouth=0.2
-    costtype='scoreiou'
-    args.do_3d_plot = True
     min_len = 4
-    
+    all_paths = {}
     for videoname in video_list:
         total_dets = 0
         video_dir = args.det_save_dir + videoname + '/'
         assert os.path.isdir(
             video_dir), 'Detection should exist @ ' + video_dir
-
-        frames_names = os.listdir(video_dir)
-        frame_ids = [int(fn.split('.')[0]) for fn in frames_names if fn.endswith('.npy')]
-        num_agent = args.num_agent
-        t1 = time.perf_counter()
-        
         agent_path_save_name = args.det_save_dir + videoname + '_agent_paths.pkl'
-        
         if args.compute_paths or not os.path.isfile(agent_path_save_name):
+            frames_names = os.listdir(video_dir)
+            frame_ids = [int(fn.split('.')[0]) for fn in frames_names if fn.endswith('.npy')]
+            num_agent = args.num_agent
+            t1 = time.perf_counter()
+            
+            
+        
             live_paths = []
             dead_paths = []
             for frame_num in sorted(frame_ids):
@@ -114,8 +111,8 @@ def perform_building(args, video_list, iteration, tube_file):
 
                     live_paths, dead_paths = gen_paths.update_agent_paths(live_paths, 
                                             dead_paths, frame, num_agent, 
-                                            iouth=iouth, time_stamp=frame_num, costtype=costtype, 
-                                            jumpgap=jumpgap)
+                                            iouth=args.paths_iouth, time_stamp=frame_num, costtype=args.paths_costtype,
+                                            jumpgap=args.paths_jumpgap)
                     
                     total_dets += num_dets
 
@@ -124,35 +121,41 @@ def perform_building(args, video_list, iteration, tube_file):
                     t1 = time.perf_counter()
             
             paths = gen_paths.copy_live_to_dead(live_paths, dead_paths)
-            paths = gen_paths.fill_gaps(paths, min_len_with_gaps=min_len*1.6)
+            paths = gen_paths.fill_gaps(
+                paths, min_len_with_gaps=args.paths_minlen, 
+                minscore=args.paths_minscore)
             ## dump agent paths to disk
             with open(agent_path_save_name,'wb') as f:
                 pickle.dump(paths, f)
         else:
             with open(agent_path_save_name, 'rb') as f:
                 paths = pickle.load(f)
-
-    return paths
-
-
-def make_tubes(args, paths, tube_file, min_len=4):
+        all_paths[videoname] = paths
     
-    detection_tubes = {}
-    for ltype in args.label_types:
-        detection_tubes[ltype] = {}
-    print('building agent tubes')
-    start_id = 1
-    agrs_vars = vars(args)
-    for ltype in args.label_types:
-            print('building tubes for ', ltype)
-            numc = agrs_vars['num_' + ltype]
-            detection_tubes[ltype][videoname] = trim_tubes(
-                start_id, numc, paths, topk=3, score_thresh=0.1, alpha=3, min_len=min_len)
-            start_id += numc
-            print(len(detection_tubes[ltype][videoname]), ltype+' tubes built')
+    return all_paths
 
-    with open(tube_file, 'w') as f:
-            json.dump(detection_tubes, f)
+
+def make_tubes(args, paths, video_list, tube_file, min_len=4):
+    """Make tubes from paths and dump in tube_file"""
+    if args.compute_tubes or not os.path.isfile(tube_file):
+        print('building agent tubes')
+        detection_tubes = {}
+        for ltype in args.label_types:
+            detection_tubes[ltype] = {}
+        for videoname in video_list:
+            start_id = 1
+            agrs_vars = vars(args)
+            for ltype in args.label_types:
+                    print('building tubes for ', ltype)
+                    numc = agrs_vars['num_' + ltype]
+                    detection_tubes[ltype][videoname] = trim_tubes(
+                        start_id, numc, paths[videoname], topk=3, alpha=3)
+                    start_id += numc
+                    print(len(detection_tubes[ltype][videoname]), ltype+' tubes built')
+
+        with open(tube_file, 'w') as f:
+                json.dump(detection_tubes, f)
+
 
 # if __name__ == '__main__':
     # main()
